+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Fix django-recaptcha tests after upgraded to 1.5.0 - 3835 by @jxltom
 - Resort imports by upgraded isort - #3836 by @jxltom
 - Support is_default_shipping/billing_address for address API - #3787 by @jxltom
+- Add bulk delete mutations - #3838 by @michaljelonek
 - Fix product creating - #3837 by @dominik-zeglen